Add notification system
When a user visits the website, a Sidekiq worker will be created if the
service that the user is trying to access (e.g. AWDWR or a Travis
status) is failing. This worker will then check to see if a notification
has been recently posted to the Campfire chat room, and if not, it will
proceed to create a Notification, and then to alert the Campfire room.
Notifications are stored using a very small, lightweight MongoDB
database, with only one model, `Notification`.